+++ src/core/meshcop/meshcop.cpp
@@ -34,7 +34,6 @@
 
 #include "meshcop.hpp"
 
-#include "common/crc.hpp"
 #include "instance/instance.hpp"
 
 namespace ot {
@@ -184,164 +183,6 @@ JoinerDiscerner::InfoString JoinerDiscerner::ToString(void) const
     }
 
     string.Append("/len:%d", mLength);
-
-    return string;
-}
-
-//---------------------------------------------------------------------------------------------------------------------
-// SteeringData
-
-Error SteeringData::Init(uint8_t aLength)
-{
-    Error error = kErrorNone;
-
-    VerifyOrExit(IsValueInRange(aLength, kMinLength, kMaxLength), error = kErrorInvalidArgs);
-
-    mLength = aLength;
-    ClearAllBytes(m8);
-
-exit:
-    return error;
-}
-
-Error SteeringData::Init(uint8_t aLength, const uint8_t *aData)
-{
-    Error error;
-
-    SuccessOrExit(error = Init(aLength));
-    memcpy(m8, aData, mLength);
-
-exit:
-    return error;
-}
-
-void SteeringData::SetToPermitAllJoiners(void)
-{
-    IgnoreError(Init(1));
-    m8[0] = kPermitAll;
-}
-
-Error SteeringData::UpdateBloomFilter(const Mac::ExtAddress &aJoinerId)
-{
-    HashBitIndexes indexes;
-
-    CalculateHashBitIndexes(aJoinerId, indexes);
-
-    return UpdateBloomFilter(indexes);
-}
-
-Error SteeringData::UpdateBloomFilter(const JoinerDiscerner &aDiscerner)
-{
-    HashBitIndexes indexes;
-
-    CalculateHashBitIndexes(aDiscerner, indexes);
-
-    return UpdateBloomFilter(indexes);
-}
-
-Error SteeringData::UpdateBloomFilter(const HashBitIndexes &aIndexes)
-{
-    Error error = kErrorNone;
-
-    VerifyOrExit(IsValid(), error = kErrorInvalidArgs);
-
-    SetBit(aIndexes.mIndex[0] % GetNumBits());
-    SetBit(aIndexes.mIndex[1] % GetNumBits());
-
-exit:
-    return error;
-}
-
-Error SteeringData::MergeBloomFilterWith(const SteeringData &aOther)
-{
-    Error error = kErrorNone;
-
-    VerifyOrExit(IsValid(), error = kErrorInvalidArgs);
-    VerifyOrExit(aOther.IsValid(), error = kErrorInvalidArgs);
-
-    VerifyOrExit(GetLength() % aOther.GetLength() == 0, error = kErrorInvalidArgs);
-
-    for (uint8_t index = 0; index < GetLength(); index++)
-    {
-        m8[index] |= aOther.m8[index % aOther.GetLength()];
-    }
-
-exit:
-    return error;
-}
-
-bool SteeringData::Contains(const Mac::ExtAddress &aJoinerId) const
-{
-    HashBitIndexes indexes;
-
-    CalculateHashBitIndexes(aJoinerId, indexes);
-
-    return Contains(indexes);
-}
-
-bool SteeringData::Contains(const JoinerDiscerner &aDiscerner) const
-{
-    HashBitIndexes indexes;
-
-    CalculateHashBitIndexes(aDiscerner, indexes);
-
-    return Contains(indexes);
-}
-
-bool SteeringData::Contains(const HashBitIndexes &aIndexes) const
-{
-    bool contains = false;
-
-    VerifyOrExit(IsValid());
-    contains = GetBit(aIndexes.mIndex[0] % GetNumBits()) && GetBit(aIndexes.mIndex[1] % GetNumBits());
-
-exit:
-    return contains;
-}
-
-void SteeringData::CalculateHashBitIndexes(const Mac::ExtAddress &aJoinerId, HashBitIndexes &aIndexes)
-{
-    aIndexes.mIndex[0] = CrcCalculator<uint16_t>(kCrc16CcittPolynomial).Feed(aJoinerId);
-    aIndexes.mIndex[1] = CrcCalculator<uint16_t>(kCrc16AnsiPolynomial).Feed(aJoinerId);
-}
-
-void SteeringData::CalculateHashBitIndexes(const JoinerDiscerner &aDiscerner, HashBitIndexes &aIndexes)
-{
-    Mac::ExtAddress address;
-
-    address.Clear();
-    aDiscerner.CopyTo(address);
-
-    CalculateHashBitIndexes(address, aIndexes);
-}
-
-bool SteeringData::DoesAllMatch(uint8_t aMatch) const
-{
-    bool matches = false;
-
-    VerifyOrExit(IsValid());
-
-    for (uint8_t i = 0; i < mLength; i++)
-    {
-        if (m8[i] != aMatch)
-        {
-            ExitNow();
-        }
-    }
-
-    matches = true;
-
-exit:
-    return matches;
-}
-
-SteeringData::InfoString SteeringData::ToString(void) const
-{
-    InfoString string;
-
-    string.Append("[");
-    string.AppendHexBytes(GetData(), Min(GetLength(), kMaxLength));
-    string.Append("]");
 
     return string;
 }
